added: exceptions for http requests

+++ src/main/java/carrico/dev/services/ExpenseServiceImpl.java
@@ -2,6 +2,8 @@ package carrico.dev.services;
 
 import carrico.dev.daos.ExpenseDAO;
 import carrico.dev.entities.Expense;
+import carrico.dev.exceptions.BadFormatException;
+import carrico.dev.exceptions.ExpenseNotFoundException;
 
 import java.util.Set;
 
@@ -13,6 +15,7 @@ public class ExpenseServiceImpl implements ExpenseService {
         this.edao = edao;
     }
 
+
     @Override
     public Expense createExpense(Expense expense) {
         long currentTime = System.currentTimeMillis();
@@ -22,18 +25,28 @@ public class ExpenseServiceImpl implements ExpenseService {
         expense.setStatus(status);
 
         expense = edao.createExpense(expense);
+        if (expense == null){
+            throw new BadFormatException();
+        }
         return expense;
     }
 
+
     @Override
     public Expense getExpenseById(int expenseId) {
-        return edao.getExpenseById(expenseId);
+        Expense e = edao.getExpenseById(expenseId);
+        if (e == null){
+            throw new ExpenseNotFoundException(expenseId);
+        }
+        return e;
     }
+
 
     @Override
     public Set<Expense> getAllExpenses() {
         return edao.getAllExpenses();
     }
+
 
     @Override
     public Expense approveExpense(int expenseId, int managerId, String reason) {
@@ -45,8 +58,14 @@ public class ExpenseServiceImpl implements ExpenseService {
         expense.setManagerReason(reason);
         expense.setStatus("approved");
 
-        return edao.updateExpense(expense);
+        Expense result = edao.updateExpense(expense);
+        if (result == null){
+            throw new BadFormatException() ;
+        }
+
+        return result;
     }
+
 
     @Override
     public Expense denyExpense(int expenseId, int managerId, String reason) {
@@ -58,11 +77,21 @@ public class ExpenseServiceImpl implements ExpenseService {
         expense.setManagerReason(reason);
         expense.setStatus("denied");
 
-        return edao.updateExpense(expense);
+        Expense result = edao.updateExpense(expense);
+        if (result == null){
+            throw new BadFormatException() ;
+        }
+
+        return result;
     }
+
 
     @Override
     public boolean deleteExpense(int expenseId) {
+        Expense e = edao.getExpenseById(expenseId);
+        if (e == null){
+            throw new ExpenseNotFoundException(expenseId);
+        }
         return edao.deleteExpenseById(expenseId);
     }
